requirements.txt
numpy>=1.24.0
matplotlib>=3.7.0
pyyaml>=6.0
pytest>=7.3.0
pytest-cov>=4.0.0
